backend/requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
motor>=3.7.0,<4.0.0
pymongo>=4.14.0,<5.0.0
pydantic-settings>=2.10.0,<3.0.0
python-jose[cryptography]>=3.5.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.1,<4.1.0
python-multipart>=0.0.20,<1.0.0
email-validator>=2.2.0,<3.0.0
pytest>=8.4.0,<9.0.0
pytest-asyncio>=1.1.0,<2.0.0
pytest-cov>=6.2.0,<7.0.0
